refactor: move deck-start filter ownership to page
