+++ src/pages/deck-start/model/useDeckFilterState.ts
@@ -1,5 +1,5 @@
 /**
- * @file Provides the study feature's Deck filter state hook.
+ * @file Provides the deck-start page's Deck filter state hook.
  * The hook combines state and operations behind one interface so components do not need to
  * coordinate services themselves.
  */
@@ -11,7 +11,7 @@ import { useForm, useWatch } from "react-hook-form";
 
 import { useAuthSession } from "@/entities/auth";
 import { editDeck } from "@/entities/deck";
-import type { DeckStartForm } from "../components/DeckStartForm";
+import type { DeckStartForm } from "../ui/DeckStartForm";
 
 type DeckStartFormProps = React.ComponentProps<typeof DeckStartForm>;
 
@@ -20,9 +20,6 @@ export interface UseDeckFilterStateOptions {
   tags: string[];
 }
 
-/**
- * Provides the filter state and persistence callback used to configure a study session.
- */
 export const useDeckFilterState = ({ deck, tags }: UseDeckFilterStateOptions): DeckStartFormProps => {
   const auth = useAuthSession();
   const uid = auth.status === "authenticated" ? auth.uid : "";
