+++ src/pages/deck-start/ui/TagFilter.tsx
@@ -1,5 +1,5 @@
 /**
- * @file Defines the study feature's Tag Filter presentation component.
+ * @file Defines the deck-start page's Tag Filter presentation component.
  * The component renders props and reports user intent through callbacks while data access stays
  * outside the view.
  */
@@ -10,10 +10,6 @@ import { Button } from "@/shared/ui/button";
 import { TagList } from "@/shared/ui/content";
 import { Switch, Tag } from "@/shared/ui/forms";
 
-/**
- * Toggles one tag in the current selection without mutating the original array.
- * Selecting an existing tag removes it; selecting a new tag appends it for the filter callback.
- */
 const updateTags = (tags: string[], tag: string) => {
   if (tags.includes(tag)) {
     return tags.filter((t) => t !== tag);
@@ -33,11 +29,6 @@ export interface TagFilterProps {
   scroll?: boolean;
 }
 
-/**
- * Renders the Tag Filter user interface.
- * Lets the user choose filter mode and tag selections, then reports the resulting filter or a
- * clear action.
- */
 export const TagFilter: React.FC<TagFilterProps> = (props) => {
   const idPrefix = useId();
   const headingId = `${idPrefix}-tags-heading`;
